requirements.txt
Flask==2.3.2
flask_restful==0.3.10
PyJWT==2.8.0
